+++ src/cli/src/commands/watch.ts
@@ -11,14 +11,12 @@
  * wrapper to enable true hot-reloading without game restarts.
  */
 
-import { type Dirent, type FSWatcher, type Stats, watch, type WatchListener, type WatchOptions } from "node:fs";
+import { type FSWatcher, type Stats, watch, type WatchListener, type WatchOptions } from "node:fs";
 import { readdir, readFile, stat } from "node:fs/promises";
 import path from "node:path";
 import process from "node:process";
-import { setImmediate as scheduleImmediate, setTimeout as scheduleTimeout } from "node:timers";
 
 import { Core, type DebouncedFunction } from "@gmloop/core";
-import { Parser } from "@gmloop/parser";
 import type * as TranspilerTypes from "@gmloop/transpiler";
 import { Transpiler } from "@gmloop/transpiler";
 import { Command, Option } from "commander";
@@ -76,7 +74,6 @@ import {
     DEFAULT_TRANSIENT_EMPTY_FILE_READ_RETRY_COUNT,
     DEFAULT_TRANSIENT_EMPTY_FILE_READ_RETRY_DELAY_MS,
     DEFAULT_WATCH_DEBOUNCE_DELAY_MS,
-    DEFAULT_WATCH_IGNORED_DIRECTORY_NAMES,
     DEFAULT_WATCH_MAX_CONCURRENT_DIRS,
     DEFAULT_WATCH_MAX_PATCH_HISTORY,
     DEFAULT_WATCH_POLLING_INTERVAL_MS,
@@ -90,6 +87,15 @@ import {
     removeDeletedCachedPatchSources,
     retranspileDependentFiles
 } from "./watch/dependency-updates.js";
+import {
+    collectScriptNames,
+    collectWatchedFilePaths,
+    EVENT_LOOP_YIELD_INTERVAL,
+    MAX_CONCURRENT_STARTUP_FILES,
+    partitionScannedDirectoryEntries,
+    shouldIgnoreWatchedPath,
+    yieldToEventLoop
+} from "./watch/directory-scan.js";
 import { handleResourceFileChange, primeRoomResource } from "./watch/resource-change-handler.js";
 import {
     clearInitialFileDataCache,
@@ -98,7 +104,6 @@ import {
     createExtensionMatcher,
     ensureScriptNameRegistered,
     type ExtensionMatcher,
-    getScriptNameFromPath,
     hashSourceContent,
     type InitialFileData,
     readSourceFileWithTransientEmptyRetry,
@@ -107,41 +112,7 @@ import {
 } from "./watch/source-analysis.js";
 
 const { debounce, getErrorMessage, isErrorWithCode } = Core;
-const IGNORED_WATCH_DIRECTORY_NAMES = new Set(DEFAULT_WATCH_IGNORED_DIRECTORY_NAMES);
-
-// ANTLR can allocate hundreds of megabytes while parsing CannonFather-sized
-// resources. Keep parser-heavy startup work tightly bounded so directory
-// traversal parallelism cannot multiply that peak beyond the process memory
-// budget. Two workers preserve useful throughput without recreating the
-// unbounded parser concurrency that exhausted large projects.
-const MAX_CONCURRENT_STARTUP_FILES = 2;
 type RuntimeDescriptorFormatter = (source: RuntimeSourceDescriptor) => string;
-
-// Startup directory/file scans call `yieldToEventLoop` many thousands of
-// times for large GameMaker projects (once or twice per file, plus once per
-// directory). A real yield costs at least ~1ms (a `setImmediate` tick plus a
-// 1ms timer, chosen so the event loop's poll phase genuinely runs before
-// resuming), so paying that cost on every call serializes into seconds of
-// pure sleep for projects with thousands of files, directly inflating watch
-// startup latency. Only every Nth call performs the real yield; the rest
-// resolve immediately. This still bounds how long CPU-heavy ANTLR parsing
-// can run before ceding control to the event loop (I/O, the WebSocket
-// server, etc.), just less often than on every single file.
-const EVENT_LOOP_YIELD_INTERVAL = 8;
-let eventLoopYieldCounter = 0;
-
-function yieldToEventLoop(): Promise<void> {
-    eventLoopYieldCounter += 1;
-    if (eventLoopYieldCounter % EVENT_LOOP_YIELD_INTERVAL !== 0) {
-        return Promise.resolve();
-    }
-
-    return new Promise((resolve) => {
-        scheduleImmediate(() => {
-            scheduleTimeout(resolve, 1);
-        });
-    });
-}
 
 /** Exposed for tests only; not part of the command's public surface. */
 export const __watchTest__ = Object.freeze({
@@ -403,19 +374,6 @@ interface FileChangeOptions extends LoggingConfig {
     abortSignal?: AbortSignal;
     /** Wall-clock timestamp (Date.now()) when the filesystem change event was first detected. */
     fileChangeDetectedAt?: number;
-}
-
-function normalizeWatchedPathSegments(candidatePath: string): Array<string> {
-    return candidatePath
-        .replaceAll("\\", "/")
-        .split("/")
-        .filter((segment) => segment.length > 0);
-}
-
-function shouldIgnoreWatchedPath(candidatePath: string, watchRoot: string | null = null): boolean {
-    const pathToCheck = watchRoot === null ? candidatePath : path.relative(watchRoot, candidatePath);
-
-    return normalizeWatchedPathSegments(pathToCheck).some((segment) => IGNORED_WATCH_DIRECTORY_NAMES.has(segment));
 }
 
 async function runAutoInjectHotReload(
@@ -2023,207 +1981,4 @@ async function updateFileSnapshot(runtimeContext: FileSnapshotWriter, filePath: 
     } catch {
         runtimeContext.fileSnapshots.delete(filePath);
     }
-}
-
-/**
- * Return value from the initial file cache build step.
- * Provides both the complete set of known script names (for seeding the semantic oracle)
- * and the per-file source content + metadata cache used between startup passes.
- * The cache intentionally excludes ASTs so large projects do not retain every parsed
- * tree until the watcher is ready.
- * `secondaryFilePaths` carries paths discovered for an additional extension matcher during the
- * same tree walk (for example `.yy` room resources) so callers do not need a second full scan.
- */
-interface InitialFileScanResult {
-    scriptNames: Set<string>;
-    fileDataCache: Map<string, InitialFileData>;
-    macroDefinitionsBySourcePath: TranspilerTypes.MacroDefinitionsBySourcePath;
-    secondaryFilePaths: Array<string>;
-}
-
-interface ScannedDirectoryEntries {
-    files: Array<string>;
-    directories: Array<string>;
-    secondaryFiles: Array<string>;
-}
-
-function partitionScannedDirectoryEntries(
-    currentPath: string,
-    entries: Array<Dirent>,
-    extensionMatcher: ExtensionMatcher,
-    watchRoot: string,
-    secondaryExtensionMatcher?: ExtensionMatcher
-): ScannedDirectoryEntries {
-    const files: Array<string> = [];
-    const directories: Array<string> = [];
-    const secondaryFiles: Array<string> = [];
-
-    for (const entry of entries) {
-        const candidatePath = path.join(currentPath, entry.name);
-        if (entry.isDirectory()) {
-            if (IGNORED_WATCH_DIRECTORY_NAMES.has(entry.name)) {
-                continue;
-            }
-            directories.push(candidatePath);
-        } else if (entry.isFile() && !shouldIgnoreWatchedPath(candidatePath, watchRoot)) {
-            if (extensionMatcher.matches(entry.name)) {
-                files.push(candidatePath);
-            } else if (secondaryExtensionMatcher?.matches(entry.name) && isRoomResourcePath(candidatePath)) {
-                secondaryFiles.push(candidatePath);
-            }
-        }
-    }
-
-    return { files, directories, secondaryFiles };
-}
-
-function isRoomResourcePath(filePath: string): boolean {
-    return path.normalize(filePath).split(path.sep).includes("rooms");
-}
-
-async function collectScriptNames(
-    rootPath: string,
-    extensionMatcher: ExtensionMatcher,
-    maxConcurrentDirs: number,
-    secondaryExtensionMatcher?: ExtensionMatcher
-): Promise<InitialFileScanResult> {
-    const scriptNames = new Set<string>();
-    const fileDataCache = new Map<string, InitialFileData>();
-    const macroDefinitionsBySourcePath: TranspilerTypes.MacroDefinitionsBySourcePath = new Map();
-    const secondaryFilePaths: Array<string> = [];
-
-    let pendingDirectories: Array<string> = [rootPath];
-    while (pendingDirectories.length > 0) {
-        const currentBatch = pendingDirectories;
-        pendingDirectories = [];
-
-        // Each iteration processes one BFS level of the directory tree; the next
-        // level's `pendingDirectories` isn't known until this level's scan finishes,
-        // so the await is a genuine sequential dependency, not an accidental serialization.
-        // eslint-disable-next-line no-await-in-loop -- level-by-level BFS traversal; see comment above
-        await Core.runInParallelWithLimit(
-            currentBatch,
-            async (currentPath) => {
-                await yieldToEventLoop();
-
-                try {
-                    const entries = await readdir(currentPath, { withFileTypes: true });
-                    const { files, directories, secondaryFiles } = partitionScannedDirectoryEntries(
-                        currentPath,
-                        entries,
-                        extensionMatcher,
-                        rootPath,
-                        secondaryExtensionMatcher
-                    );
-
-                    // Keep lexer and source-directive work globally bounded by
-                    // the same worker limit as directory traversal.
-                    for (const filePath of files) {
-                        // eslint-disable-next-line no-await-in-loop -- intentionally serialized to respect maxConcurrentDirs; see comment above
-                        await addScriptNamesFromFile(
-                            filePath,
-                            scriptNames,
-                            fileDataCache,
-                            macroDefinitionsBySourcePath
-                        );
-                    }
-
-                    // Capture paths discovered for the secondary matcher without
-                    // re-walking the tree during watch startup.
-                    secondaryFilePaths.push(...secondaryFiles);
-                    pendingDirectories.push(...directories);
-                } catch {
-                    // Ignore per-directory read errors; the watcher can still use
-                    // file-name fallback and process later change events.
-                }
-            },
-            Math.min(maxConcurrentDirs, MAX_CONCURRENT_STARTUP_FILES)
-        );
-    }
-
-    return { scriptNames, fileDataCache, macroDefinitionsBySourcePath, secondaryFilePaths };
-}
-
-async function collectWatchedFilePaths(
-    rootPath: string,
-    extensionMatcher: ExtensionMatcher,
-    maxConcurrentDirs: number
-): Promise<Array<string>> {
-    const discoveredFiles: Array<string> = [];
-
-    async function scan(currentPath: string): Promise<void> {
-        try {
-            await yieldToEventLoop();
-            const entries = await readdir(currentPath, { withFileTypes: true });
-            const { files, directories } = partitionScannedDirectoryEntries(
-                currentPath,
-                entries,
-                extensionMatcher,
-                rootPath
-            );
-
-            discoveredFiles.push(...files);
-
-            await Core.runInParallelWithLimit(
-                directories,
-                async (subDirPath) => {
-                    await scan(subDirPath);
-                },
-                maxConcurrentDirs
-            );
-        } catch {
-            // Ignore per-directory read errors; the unknown scan should never
-            // crash the watcher just because one subdirectory is inaccessible.
-        }
-    }
-
-    try {
-        await scan(rootPath);
-    } catch {
-        // Fail silently; unknown filename scans should never crash the watcher.
-    }
-
-    return discoveredFiles;
-}
-
-async function addScriptNamesFromFile(
-    filePath: string,
-    scriptNames: Set<string>,
-    fileDataCache: Map<string, InitialFileData>,
-    macroDefinitionsBySourcePath: TranspilerTypes.MacroDefinitionsBySourcePath
-): Promise<void> {
-    const beforeSize = scriptNames.size;
-
-    try {
-        // `readFile` and `stat` are independent I/O operations; pipeline them so
-        // the per-file startup cost tracks the slower of the two instead of the
-        // sum. The pipelined `try` still catches a stat failure alongside the
-        // read failure, preserving the original "ignore parse errors" fallback
-        // behavior used by the directory walk.
-        const [content, stats] = await Promise.all([readFile(filePath, "utf8"), stat(filePath)]);
-        await yieldToEventLoop();
-        for (const functionName of Parser.extractGmlFunctionNames(content)) {
-            scriptNames.add(`gml_Script_${functionName}`);
-        }
-
-        if (sourceCanDeclareMacroMetadata(content)) {
-            macroDefinitionsBySourcePath.set(filePath, Transpiler.extractMacroDefinitionsFromSource(content, filePath));
-        }
-
-        fileDataCache.set(filePath, { content, mtimeMs: stats.mtimeMs, symbols: [], references: [] });
-        await yieldToEventLoop();
-    } catch {
-        // Ignore parse errors; fallback to file-name based script
-    }
-
-    if (scriptNames.size === beforeSize) {
-        const scriptName = getScriptNameFromPath(filePath);
-        if (scriptName) {
-            scriptNames.add(scriptName);
-        }
-    }
-}
-
-function sourceCanDeclareMacroMetadata(content: string): boolean {
-    return content.includes("#macro") || content.includes("#define");
 }
