+++ src/cli/src/commands/watch/index.ts
@@ -1,4 +1,5 @@
 export * from "./constants.js";
+export * from "./directory-scan.js";
 export * from "./resource-change-handler.js";
 export * from "./source-analysis.js";
 export * from "./status.js";
